+++ data/robots.sql
@@ -1,0 +1,132 @@
+-- phpMyAdmin SQL Dump
+-- version 4.6.6
+-- https://www.phpmyadmin.net/
+--
+-- Host: localhost
+-- Generation Time: Mar 31, 2017 at 12:04 PM
+-- Server version: 5.6.35
+-- PHP Version: 5.6.29
+
+SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
+SET time_zone = "+00:00";
+
+
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
+/*!40101 SET NAMES utf8mb4 */;
+
+--
+-- Database: `robots`
+--
+
+-- --------------------------------------------------------
+
+--
+-- Table structure for table `parts`
+--
+
+CREATE TABLE `parts` (
+  `certificate` char(6) NOT NULL,
+  `model` enum('a','b','c','m','r','w') NOT NULL,
+  `piece` tinyint(1) NOT NULL,
+  `plant` varchar(32) NOT NULL,
+  `available` tinyint(1) NOT NULL DEFAULT '1',
+  `built` datetime NOT NULL
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+-- --------------------------------------------------------
+
+--
+-- Table structure for table `properties`
+--
+
+CREATE TABLE `properties` (
+  `property_name` varchar(16) NOT NULL,
+  `value` varchar(32) NOT NULL
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+--
+-- Dumping data for table `properties`
+--
+
+INSERT INTO `properties` (`property_name`, `value`) VALUES
+('api_key', ''),
+('balance', ''),
+('server_url', 'http://umbrella.jlparry.com/'),
+('team', 'Nectarine');
+
+-- --------------------------------------------------------
+
+--
+-- Table structure for table `robots`
+--
+
+CREATE TABLE `robots` (
+  `id` int(11) NOT NULL,
+  `head` char(6) NOT NULL,
+  `torso` char(6) NOT NULL,
+  `legs` char(6) NOT NULL,
+  `built` datetime NOT NULL
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+-- --------------------------------------------------------
+
+--
+-- Table structure for table `transactions`
+--
+
+CREATE TABLE `transactions` (
+  `id` int(11) NOT NULL,
+  `type` enum('part_sale','part_purchase','robot_sale','robot_purchase','part_recycle','part_build','box_purchase','robot_build') NOT NULL,
+  `part_id` char(6) NOT NULL,
+  `robot_id` int(11) NOT NULL,
+  `amount` int(11) NOT NULL,
+  `time` datetime NOT NULL
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+--
+-- Indexes for dumped tables
+--
+
+--
+-- Indexes for table `parts`
+--
+ALTER TABLE `parts`
+  ADD PRIMARY KEY (`certificate`);
+
+--
+-- Indexes for table `properties`
+--
+ALTER TABLE `properties`
+  ADD PRIMARY KEY (`property_name`);
+
+--
+-- Indexes for table `robots`
+--
+ALTER TABLE `robots`
+  ADD PRIMARY KEY (`id`);
+
+--
+-- Indexes for table `transactions`
+--
+ALTER TABLE `transactions`
+  ADD PRIMARY KEY (`id`);
+
+--
+-- AUTO_INCREMENT for dumped tables
+--
+
+--
+-- AUTO_INCREMENT for table `robots`
+--
+ALTER TABLE `robots`
+  MODIFY `id` int(11) NOT NULL AUTO_INCREMENT;
+--
+-- AUTO_INCREMENT for table `transactions`
+--
+ALTER TABLE `transactions`
+  MODIFY `id` int(11) NOT NULL AUTO_INCREMENT;
+/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
+/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
+/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;